requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain-google-genai
langchain
streamlit
langchain_community
python_dotenv
pypdf